refactor: Review score 타입 변경 #2

+++ src/main/java/com/yamyamnavi/api/v1/response/ReviewResponse.java
@@ -9,7 +9,7 @@ public record ReviewResponse (
 
     @Schema(description = "아이디") Long id,
     @Schema(description = "내용") String content,
-    @Schema(description = "평점") Double score,
+    @Schema(description = "평점") Integer score,
     @Schema(description = "작성자 아이디") Long userId,
     @Schema(description = "가게 아이디") Long restaurantId,
     @Schema(description = "작성 일시") LocalDateTime createdAt
